+++ R/filterResultsLI.R
@@ -537,6 +537,7 @@ FilterResultsLI <- setRefClass(
         cat("\n")
         cat("Seasonality noise:", format(Q_seasonal, digits = 4))
       }
+      cat("\n")
     },
     mapes=function(n.ahead,Y){
       "Computes five metrics, including Mean Absolute Percentage Error (MAPE),